pyproject.toml
[tool.poetry]
name = "letterboxd-list"
version = "0.1.0"
description = ""
authors = ["Hugo Delahousse <hugo.delahousse@epita.fr>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.1"
requests_html = "^0.10.0"
PyGithub = "^1.44.1"
dataclasses_json = "^0.3.7"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
letterboxd-to-gist = "letterboxd_list.gist_handler:main"
